Reconstruct of branch tech_soil. All changes of the branch in this one huge commit.
For the history of the individual commits done during the developments in the original branch, see the log of the branch tech_soil: r4794 - r7446.

The original branch might in the future be removed from HEAD, but click on the revisions here and you will find it.

The reason for reconstruction was corrupted svn:properties.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/arctic_and_wetlands@7449 70805675-57b6-444b-bb00-18ecffcb4be8