suppressed warnings
Initialized root_beta to 0 to suppress valgrind warnings

Implemented simple verbosity level to suppress warnings from methane module. This lets the user decide the amount of information that it wants from the model. Default is to show errors and warnings but not general information or debug messages. Can be extended to other code later on.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/arctic_and_wetlands_bugfix@7648 70805675-57b6-444b-bb00-18ecffcb4be8