Fixed that randfrac() gave different results and seeds on windows vs linux.
The hypothesis for origin of difference was different bit length of long on the 2 systems.

Solution:
A fixed bit lenght datatype is needed locally in randfrac() function to get identical seed numbers on windows machines as on linux ones.
However, the seed is never larger than the largest number stored by long on windows (2147483647, i.e. constant modulus), 
so the value of seed can safely be stored in long seed between calls to randfrac().

Test: windows VSc2013 and simba give same debug output of seed and randfrac (debug dprinf in randfrac()).

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_fixes@7861 70805675-57b6-444b-bb00-18ecffcb4be8