Fixed bug that input-module=emdi somehow stayed in submit_vars.sh for emdi BMs despite apparent sync with trunk.
The present branch was synced with trunk getting 7458 (merginfo: /LPJ-GUESS/trunk:5986-6961,6964-7143-->6964-7484, so getting 7458)
which is arctic and wetlands feature branch.

This sync deleted benchmarks/extra_source/emdi
and benchmarks/emdi_global/extra_source.txt

but for some unclear reason, the change to trunk (r7458) in the following files were not transfered to the current branch!
/LPJ-GUESS/trunk/benchmarks/emdi_europe/submit_vars.sh
/LPJ-GUESS/trunk/benchmarks/emdi_global/submit_vars.sh

This missing change is fixed in this commit.

(benchmarks/emdi_europe/extra_source.txt were also not deleted but makes sense, since it was deleted from trunk in a later A&W bugfix )

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/fire_blaze_merge@7891 70805675-57b6-444b-bb00-18ecffcb4be8