Reverted all changes to aurora.tmpl. These changes will be added after 4.1-code-freeze instead from a aurora/sumbit-script dedicated branch.
Reason. These changes are not blaze related and the aurora.tmpl-changes are not mature enough to go in to trunk, and other changes to aurora.tmpl/sumbit-script are planned anyway.

E.g. an additional change that need to be investigated: it seems as 
--bind-to core 
needs to be conditional depending on whether the EXCLUSIVE flag is set or not. But we need to look into this more.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/fire_blaze_merge@8208 70805675-57b6-444b-bb00-18ecffcb4be8