Fusing all 3 main ins-files in to one general one. Step 2a: moving group 'common' to general settings.
Step 2 is: Fusing common groups and possibly common Pfts

What is remaining of step 2 is to tend to some groups that are duplicated with small differences between the duplicates.
E.g. tree

Stefan has a good solution for this: create new group european_tree which includes tree (common parts of tree).


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_fixes@8259 70805675-57b6-444b-bb00-18ecffcb4be8