Adjustment of the new code for function describe_images and the call to it from fluxnet/postprocess.sh
I opted for a more conservative solution where the embed keyword is placed as the last argument instead of as the second.
This solution is similar to how other commands take embed. The previous variant was not.

Tested to work on benchmark_fluxnet r8653 benchmark fluxnet.
The previous commit still needs to be tested on also e.g. crop_global.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_fluxnet@8670 70805675-57b6-444b-bb00-18ecffcb4be8