Removed superfluous comments
I removed them because they break the over all style of the crop_global benchmark code.
If they turn out to be necessary, bring them back later.
I'd rather prefer to move the bulky code to a sub routine, in order to make the main code so transparent that 
comments are not necessary.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_biomass@9263 70805675-57b6-444b-bb00-18ecffcb4be8