Changed CMakeList.txt for semi-automated finding of NetCDF library, plus added new warning message from cmake when not found.
For a general system, an outcommented line for the user to modify.
for aurora: automate finding the root path to NetCDF library etc.
Also added (on Lars' request), an error message for cmake, when the library is not found by cmake.

For the general system: The change assumes that an instruction text for the user be added in guessdoc.doc (written , will be added).

The solution is based on Peter's suggestion set(CMAKE_PREFIX_PATH
in combination with my own $ENV{EBROOTNETCDF}, which automatically gets the root path to netcdf library etc on an easybulid system.

Test:
The solution was tested for Cmake 3.12 on aurora, including a negative text, which ouputed the desired error msg.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@9491 70805675-57b6-444b-bb00-18ecffcb4be8