Reintegrated branch bugfix (r9619-9620): denitrification function bug fix.
Contributor: Stefan
 
Main change:
A bug fix in the denitrification function and new parameters affecting the denitrification 
changed due to fixing the bug.

No changes to inputs nor to outputs.

Benchmarks results: because this is an single bug fix in a specific feature, no benchmark run 
was done. Potentially, its effects om global benchmark will be accounted for in the log 
comment of a later revsion of trunk.

Branch history: 
Created from trunk r9565 @ 9619.
This fix was initially committed to branch release_finetunings r9581,
and later moved to branch bugfix @9620.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@9621 70805675-57b6-444b-bb00-18ecffcb4be8