1st step to fix CMakeLists.txt's ability to find NetCDF on aurora. Buggy: does not work on simba.
On aurora: the fix should actually work on any Easybuild system with $ARCH=aurora.

This 1st step is the solution that I erroneously committed to trunk r9491.
(I used that diff to create this commit).
It's also the one that I emailed to Peter 2021-02-12.

Now tried out on simba:/scratch/johan/bms_n_dev/techfixes-br/svntechfixes.
It works on aurora but not on simba becasue of missing \"
Fixed in next commit (i.e. step 2).


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_fixes@9683 70805675-57b6-444b-bb00-18ecffcb4be8