Added possibility to specify wood cutting timings relative to firstcutyear instead of using patch age.
If firstcutyear_is_referenceyear is set to 1 in a st or mt definition in the instruction file, the reference age used for thintime values is simulation calender year - firstcutyear rather than the patch age. This can be used for specifying cuttings using exact calendar years when stand is not initialised by clear-cutting. All patches will be cut in synchrony.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/forestry_4.1.2@10320 70805675-57b6-444b-bb00-18ecffcb4be8