Correction of erroneous usage of mt.target_thinning_select_diam and mt.target_thinning_select_age
These parameters were not used correctly in the code. In the call to distribute_cutting() in set_forest((), hard-coded values of 2 and 0 (cut largest trees first) were used. Now the correct instruction file parameters are used and set by default to the same values.

Also, mt.target_thinning_select_age with the previously default value of 1 (cut young first) was used in a call to distribute_cutting() in thin_reineke(). Now, this value is hard-coded in the function call.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/forestry_4.1.2@10324 70805675-57b6-444b-bb00-18ecffcb4be8