Reorganising ManagementType storage slightly.
(crop_ncep branch r10499 & r10613)

mt:s defined directly in a stand type are now copied to an item in the mtlist with the same name as the stand type. The st.management copy is now used only to allow mt:s to be defined directly in a stand type in the instruction file (when there is no rotation between management types) and the copy in the mtlist is used after that to retrieve mt information for a stand or stand type. 

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/forestry_4.1.2@10625 70805675-57b6-444b-bb00-18ecffcb4be8