Transfer (by merging) of 5 commits in branch bugfix to this branch tech_fixes2. These commits will next be removed (reverted) from branch bugfix.
Reason: these commits (changes) do not change the model behavior, and thus need to be not in branch bugfix.

The 5 commits are listed below.
The HEAD of branch bugfix when this is done is r11269.
The trunk-sync status of branch bugfix when this is done is: 
bugfix has been synced with up to trunk r11208, i.e. all trunk revisions excep the latest one: r11300

The trunk-sync status of the destination branch tech_fixes2 (this branch) is:
tech_fixes2 has been created from trunk r11208 (no other changes), 
and thus has the same trunk-sync status as branch bugfix.


The 5 commits in branch bugfix transfered here are: 

r11269
Adding the name of the missing map file in an error message in TimeDataD::CreateFileMap(). 

r11265
Units for HYDE3.1 data corrected in the comments. Not affecting the actual code. 

r11259
Fixing spelling.

r11245
Fixing some compilation warnings on Mac.

r11234
Fix of a bug introduced in r11208 (forestry_4.1.2) causing an error on MAC systems. This fix has not yet been run.
The windows specific code line that caused the error on MAC is now hidden behind a windows-checking #ifdef.
This fix compiles on windows but has not yet been run.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_fixes2@11326 70805675-57b6-444b-bb00-18ecffcb4be8