Updated the exoprocesses/tellme/exoprocess.sh script according to instructions from Matt email 2023-03-13.
TODO: The script needs to be updated with code for determining the path to the Rscript tellme.Rmd 
which will probably, I guess, be sitting in "svnhead"/benchmarks/exoprocesses/tellme/
I e. the code needs to figure out, or get as a call argument, the path to the LPJG code folder.

Also, see r10728 for a list of TODOs.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_tellme@11712 70805675-57b6-444b-bb00-18ecffcb4be8