Fix for "out-of-mem error": added EXCLUSIVE flag to the xpp-batchjob.
Now the job passed the point previously stopped by the "out-of-mem error".

Todo: check that lunarcify adds the lunarc prefix to the data path in tellme/exoprocess.sh
Also check that the $SBATCH_ACCOUNT is added on line 310 in benchmarks/benchmarks
or rather that it does not clash with the line already added in this branch.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_tellme@11802 70805675-57b6-444b-bb00-18ecffcb4be8