Paths and variable names to cruncep monthly files for testing on simba.
The paths are temporary: the paths should be /data/somtehing in the final version.

I did not supply max and min-temp files. So I need to run with BVOCs off.
Is it correct to use the "variable_insol"?

A new ins-file last.ins introduced in benchmarks/
All BM's guess.ins are supposed to import it. For now only gl, eur, and crg imports it.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmarks_crujra_cf@12231 70805675-57b6-444b-bb00-18ecffcb4be8