Monthly paths.ins - cosmos paths and local CO2-afile. 1st of 2 in 2-step commit. Worked 240119.
This is the paths.ins that work together with the fixed cfinput.cpp of r12579.
Most of the monthly crujra files are from Kutabiri:/kukusanya/petera/ 240111
except for wetdays crujra.v2.4.5d.wet.1901-2022.365d.noc.if1mm.monsum.nc that I made 240118.

next: after bm-suit shows that the mothly data files are ok:
addapt the paths.ins into simba style,

What about the the hybrid-CO2? - place it in the /data/benchmarks_data on cosmos eventhough it is kind of temporary.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmarks_crujra_cf@12584 70805675-57b6-444b-bb00-18ecffcb4be8