Reverted r12335 (-C option: no compile). The code (which does not work) will be placed in a sep branch "benchmark_tellme_aux",
or just revived in this branch from r12335 at a later point.

The problem with the code for option -C is that it appears as -C does reverse of what it should:
builds with -C, an does not build without -C, but if echo debug-output is placed to show the if-exec-flow
then it shows that the problem is elsewhere downstream of the if [ $no_compile.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_tellme@12692 70805675-57b6-444b-bb00-18ecffcb4be8