Further fixes to make it possible to run several tellme-report-generations with one call to ./benchmarks. Tested. Works.
With this, it is possible to run several tellme-report-generations with one call to ./benchmarks
The tellme folder on the output directory is no longer erased at the start of the run,
so several tellme-report outputs can be done to the tellme folder.

The report names take this form
"Unique" filename: tellme_global.${NAMETAG_NEW}.vs.${NAMETAG_REF}.html
The exoprocess.debug.log is appended to, so all history is captured here.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_tellme@12700 70805675-57b6-444b-bb00-18ecffcb4be8