Moved revision r13186 from branch tech_various to here. Demo instruction files changed from symbolic filenames without paths,
Reason: ready for trunk.

Original log message r13186 in tech_various:
Demo instruction files changed from symbolic filenames without paths, into relative paths that work for running the model out of the box without any modification. Just build and run.

The model runs out of the box. No modification of forcing-file paths in the demo instruction files are needed.
All that is needed to start a model run is:
svn checkout <url> .
cd build 
cmake ..
make
guess -input demo ../data/ins/europe_demo.ins
or 
guess -input demo ../data/ins/global_demo.ins

Benefit for model maintenance:
This facilitates for developers when quickly fixing a bug:
- now thay can easlily run-test the model before they report that they have fixed a bug.
Because now it is as easy to run the model as it is to compile it.

the above schem is for linux, but it is the same on windows
Only difference: build in VS, and run in dos with guesscmd.exe -input demo ../data/ins/europe_demo.ins

------------------------------------------------------------------------


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/techfix@13208 70805675-57b6-444b-bb00-18ecffcb4be8