Moved r13359-13457 from branch tech_various to here: New algorithm for split_gridlist that doesnt give empty gridlists.
Tested successfully:
global benchmark (only the 1st 6000 gridcells) split onto either 40 or 80 or 71 cores.
Verified that the order of output is identical when using two different splits of the gridlist (e.g. in 40 parts and in 80 parts),
and that the gridcell order in the output is the same as the order in the gridlist.
For details see the issue tracker on gitlab update `split_gridlist` in submit templates 
 https://stormbringer4.nateko.lu.se/lpj-guess-developers/LPJ-GUESS/-/issues/20

The changes in tech_various transferred to here:
r13359 New algorithm for split_gridlist that hopefully doesnt give empty gridlists. 
r13360 New algorithm for split_gridlist: fixed errors in comments. 
r13445 Bugfix. Error "bad substitution".
r13446 Adjusted submitvars lunarc values of bms emdi_* to better values now that is possible thanks to the new split algorithm. 
r13457 Exactly same change as to simba.tmpl to the other tmpl-files.

Potential improvement:
Perhaps add a fail if the original gridlist is smaller than N_processes.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/techfix@13587 70805675-57b6-444b-bb00-18ecffcb4be8